feat: Shebang & Imports & Rename from .hcl to .variant
Add `#!/usr/bin/env variant` to any file and set the executable bit on it to turn that into an executable command.

Also Adds `import: "path/to/dir"` so that you can import any directory containing *.variant files to externalize and split definition(s) of your command. This is handy when used with the shebang. See the test/shebang example for more information.

Sorry to make this a big commit but I have also included the rename from `.hcl` to `.variant` according to the convention that any DSL build on top of HCL should use the file extension other than `.hcl`, like Terraform uses `.tf` instead of `.hcl`.