Pass globals to imported jobs
Imported variant command's globals (top-level parameters and options) are now merged into the command, when it misses parameters and options with the same names.

If any imported global's type conflict with the command's corresponding parameter/option type, it fails early so that the imported command doesn't fail due to unexpected value/type provided.

Resolves #29